+++ cpp/arduino/ArduinoDefines.h
@@ -87,4 +87,3 @@
 #define TIMER5B 17
 #define TIMER5C 18
 
-#include "include/io.h"