+++ cpp/arduino/AvrMath.h
@@ -4,6 +4,11 @@
 #define constrain(x,l,h) ((x)<(l)?(l):((x)>(h)?(h):(x)))
 #define map(x,inMin,inMax,outMin,outMax) (((x)-(inMin))*((outMax)-(outMin))/((inMax)-(inMin))+outMin)
 
+#define sq(x) ((x)*(x))
+
+#define radians(deg) ((deg)*DEG_TO_RAD)
+#define degrees(rad) ((rad)*RAD_TO_DEG)
+
 #ifdef abs
 #undef abs
 #endif
@@ -18,10 +23,4 @@
 #undef min
 #endif
 #define min(a,b) ((a)<(b)?(a):(b))
-
-#define sq(x) ((x)*(x))
-
-#define radians(deg) ((deg)*DEG_TO_RAD)
-#define degrees(rad) ((rad)*RAD_TO_DEG)
-