+++ cpp/arduino/Godmode.cpp
@@ -46,6 +46,27 @@ long random(long vmin, long vmax)
   return vmin < vmax ? (random(vmax - vmin) + vmin) : vmin;
 }
 
+void digitalWrite(unsigned char pin, unsigned char val) {
+  GodmodeState* godmode = GODMODE();
+  godmode->digitalPin[pin] = val;
+}
+
+int digitalRead(unsigned char pin) {
+  GodmodeState* godmode = GODMODE();
+  return godmode->digitalPin[pin].retrieve();
+}
+
+void analogWrite(unsigned char pin, int val) {
+  GodmodeState* godmode = GODMODE();
+  godmode->analogPin[pin] = val;
+}
+
+int analogRead(unsigned char pin) {
+  GodmodeState* godmode = GODMODE();
+  return godmode->analogPin[pin].retrieve();
+}
+
+
 // Serial ports
 #if defined(HAVE_HWSERIAL0)
   HardwareSerial Serial(&godmode.serialPort[0].dataIn, &godmode.serialPort[0].dataOut, &godmode.serialPort[0].readDelayMicros);
@@ -60,3 +81,8 @@ long random(long vmin, long vmax)
   HardwareSerial Serial3(&godmode.serialPort[3].dataIn, &godmode.serialPort[3].dataOut, &godmode.serialPort[3].readDelayMicros);
 #endif
 
+template <typename T>
+inline std::ostream& operator << ( std::ostream& out, const PinHistory<T>& ph ) {
+  out << ph;
+  return out;
+}